+++ src/ValidatorInterface.php
@@ -7,15 +7,11 @@ use Dhii\Validation\Exception\ValidationFailedExceptionInterface;
 
 /**
  * Something that can validate a value.
- *
- * @since 0.1
  */
 interface ValidatorInterface
 {
     /**
      * Validates a value.
-     *
-     * @since 0.1
      *
      * @param mixed $value The subject of validation.
      *